html/semantics/forms/the-input-element/input-type-color.html

<!DOCTYPE HTML>
<html>
 <head>
  <title> New Document </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>

 <body onload="changeVal()">
 <h1>input element type=color</h1>
 <input type="color" id="ipt" />

 <script>
 function changeVal() {
	document.getElementById("ipt").value == "";
 }

 test(function() {
	 assert_false(document.getElementById("ipt").value == "");
	 }, "the value cannot be empty string");

 </script>
 
 <div id="log"></div>
 </body>
</html>
